% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{createStanModel}
\alias{createStanModel}
\title{Create an .stan model}
\usage{
createStanModel(P = 1)
}
\arguments{
\item{P}{The highest power of WBS signal in regression}
}
\value{
The .stan model
}
\description{
This is an internal function used to create a Stan model.
}
\examples{
cat(createStanModel(3))
}
