% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\docType{data}
\name{exampledata}
\alias{exampledata}
\title{Sample data from the Calgary study}
\format{
An object of class \code{"data.table"}
}
\usage{
data(exampledata)
}
\description{
A R data table containing Date, Number of clinical cases,
Positivity rate, normalized N1 and N2 copies, and vaccination rate
}
\examples{
data(exampledata)

}
\keyword{datasets}
