% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{PoissonReg}
\alias{PoissonReg}
\title{Main Poisson/Negative Binomial regression function}
\usage{
PoissonReg(
  modeldata,
  date,
  p.cases,
  p.rate,
  ww.signal,
  lag = 1,
  start_date,
  end_date,
  StanModel,
  iteration,
  distribution = "Poisson"
)
}
\arguments{
\item{modeldata}{The long-format data frame/table containing date, WBS observation, and case count}

\item{date}{Name of date column}

\item{p.cases}{Name of case count column}

\item{p.rate}{Name of positivity rate (# of cases / # of people tested) column}

\item{ww.signal}{Names of WBS signal column}

\item{lag}{The time lag between WBS signal and case count}

\item{start_date}{Starting date of training data used to build model}

\item{end_date}{End date of training data used to build model}

\item{StanModel}{Created Stan model}

\item{iteration}{A positive integer specifying the number of iterations for each chain (including burnin).}

\item{distribution}{Distribution of outcome variable: "Poisson" or "NB".}
}
\value{
fit The fitted MCMC object
}
\description{
This is the main function for fitting Bayesian Poisson OR Negative Binomial regression model.
}
\details{
This is the main function used to fit a Bayesian regression model with Poisson-distributed response
to predict clinical cases.
See ?predictPoisson for how to make predictions of clinical cases from Bayesian regression model.
}
\examples{
modeldata = as.data.table(readRDS("modeldata.rds"))
modeldata = modeldata[modeldata$date <= as.Date("2021-12-20"), ]
modelres = PoissonReg(modeldata, date, p.cases, p.rate, ww.signal, lag,
                      start_date, end_date, iteration)
}
