% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{predictPoisson}
\alias{predictPoisson}
\title{Main prediction function}
\usage{
predictPoisson(
  modelres,
  modeldata,
  pred_start,
  pred_end,
  iteration,
  p.cases,
  p.rate,
  ww.signal
)
}
\arguments{
\item{modelres}{The model result from PoissonReg()}

\item{modeldata}{The long-format data frame/table containing date, WBS observation, and case count}

\item{pred_start}{Starting date of testing data used to make predictions}

\item{pred_end}{End date of testing data used to make predictions}

\item{iteration}{A positive integer specifying the number of iterations for each chain (including burnin).}

\item{p.cases}{Name of case count column}

\item{p.rate}{Name of positivity rate (# of cases / # of people tested) column}

\item{ww.signal}{Names of WBS signal column}
}
\value{
Ypred The forecast of observed virus concentration
}
\description{
This is the main prediction function for a fitted Poisson/Negative Binomial regression model.
}
\details{
This function is used to make predictions using a fitted Poisson regression model.
}
\examples{
\dontrun{
predres = predictPoisson(modelres, modeldata, pred_start, pred_end, iteration)
}
}
